+++ terraform/slurm_cluster/examples/slurm_cluster/cloud/winbind/main.tf
@@ -111,6 +111,7 @@ locals {
           node_conf              = {}
 
           additional_disks         = []
+          bandwidth_tier           = "platform_default"
           can_ip_forward           = false
           disable_smt              = false
           disk_auto_delete         = true
@@ -159,6 +160,7 @@ locals {
           node_conf              = {}
 
           additional_disks       = []
+          bandwidth_tier         = "platform_default"
           can_ip_forward         = false
           disable_smt            = false
           disk_auto_delete       = true