+++ terraform/slurm_cluster/examples/slurm_cluster/hybrid/full/main.tf
@@ -29,6 +29,7 @@ locals {
       partition_name            = x.partition_name
       partition_nodes = [for n in x.partition_nodes : {
         additional_disks         = n.additional_disks
+        bandwidth_tier           = n.bandwidth_tier
         can_ip_forward           = n.can_ip_forward
         node_count_dynamic_max   = n.node_count_dynamic_max
         node_count_static        = n.node_count_static