+++ terraform/slurm_cluster/examples/slurm_cluster/hybrid/simple/main.tf
@@ -35,6 +35,7 @@ locals {
           node_conf              = {}
 
           additional_disks         = []
+          bandwidth_tier           = "platform_default"
           can_ip_forward           = false
           disable_smt              = false
           disk_auto_delete         = true
@@ -85,6 +86,7 @@ locals {
           node_conf              = {}
 
           additional_disks       = []
+          bandwidth_tier         = "platform_default"
           can_ip_forward         = false
           disable_smt            = false
           disk_auto_delete       = true