+++ terraform/slurm_cluster/examples/slurm_controller_hybrid/simple/main.tf
@@ -48,6 +48,7 @@ module "slurm_partition0" {
       node_conf              = {}
 
       additional_disks         = []
+      bandwidth_tier           = "platform_default"
       can_ip_forward           = false
       disable_smt              = false
       disk_auto_delete         = true
@@ -95,6 +96,7 @@ module "slurm_partition1" {
       node_conf              = {}
 
       additional_disks       = []
+      bandwidth_tier         = "platform_default"
       can_ip_forward         = false
       disable_smt            = false
       disk_auto_delete       = true