+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## \[5.1.0\]
 
+- Add support for gvnic and tier1 networking
+
 ## \[5.0.3\]
 
 - Allow configuring controller hostname for hybrid deployments.